requirements.txt
pip>=6.0
google-cloud-translate>=1.3.1
openpyxl>=2.5.5
python-docx>=0.8.6
python-pptx0.6.12